Improve stability by not storing invalid tokens
This improves stability by not storing access tokens in the session until they have passed the verification stage. Otherwise, there is no way for the user to do anything about an invalid token that throws an exception except for wait for the token to expire and the Google API Client library to attempt to generate a new one for them.

This change makes it so that the access token is only stored in the session after successfully passing validation.